🚚 Move resistive tf calcs and cpost function into resistive file along with tests
