+++ process/resistive_tf_coil.py
@@ -1,5 +1,6 @@
 import logging
 
+import numba
 import numpy as np
 
 from process.fortran import (
@@ -14,7 +15,7 @@ from process.tfcoil import TFCoil
 
 logger = logging.getLogger(__name__)
 
-logger = logging.getLogger(__name__)
+EPS = np.finfo(1.0).eps
 
 
 class ResistiveTFCoil(TFCoil):
@@ -23,6 +24,7 @@ class ResistiveTFCoil(TFCoil):
 
     def run(self):
         self.res_tf_internal_geom()
+        self.tf_res_heating()
 
     def res_tf_internal_geom(self):
         """
@@ -159,3 +161,515 @@ class ResistiveTFCoil(TFCoil):
         elif sctfcoil_module.awptf < 0.0e0:
             error_handling.fdiags[0] = sctfcoil_module.awptf
             error_handling.report_error(101)
+
+    def tf_res_heating(self) -> None:
+        """
+        Calculate resistive heating for resistive magnets.
+
+        This method calculates the resistive heating for resistive magnets.
+        It considers the following scenarios:
+        - Clamped joints in superconductors might have resistive power losses on the joints.
+        - Sliding joints might have a region of high resistivity.
+
+        Notes:
+        - The copper resisitivty is set to be that for GLIDCOP AL-15 at 20°C for copper (i_tf_sup = 0).
+        - The coefficient of resistivity is set to be that of pure copper
+
+        References:
+            - https://www.spotweldingconsultants.com/GlidCop_AL_15.pdf
+
+            - https://cirris.com/temperature-coefficient-of-copper/
+        """
+
+        # Resistivity of the Glidcop copper centerpost
+        if tfcoil_variables.i_tf_sup == 0:
+            tfcoil_variables.rho_cp = (
+                # 1.86 is the resitivity at `20°C` for GLIDCOP AL-15
+                # 0.00393 is the coefficient of resistivity for copper
+                tfcoil_variables.frhocp
+                * (1.86e0 + 0.00393e0 * (tfcoil_variables.temp_cp_average - 293.15e0))
+                * 1.0e-8
+            )
+
+        # Resistivity of the aluminium centerpost
+        if tfcoil_variables.i_tf_sup == 2:
+            tfcoil_variables.rho_cp = tfcoil_variables.frhocp * (
+                2.00016e-14 * tfcoil_variables.temp_cp_average**3
+                - 6.75384e-13 * tfcoil_variables.temp_cp_average**2
+                + 8.89159e-12 * tfcoil_variables.temp_cp_average
+            )
+
+        # Calculations dedicated for configurations with CP
+        if physics_variables.itart == 1:
+            # Tricky trick to make the leg / CP tempearture the same
+            if (
+                abs(tfcoil_variables.temp_tf_legs_outboard + 1.0e0)
+                < np.finfo(float(tfcoil_variables.temp_tf_legs_outboard)).eps
+            ):
+                sctfcoil_module.is_leg_cp_temp_same = 1
+                tfcoil_variables.temp_tf_legs_outboard = (
+                    tfcoil_variables.temp_cp_average
+                )
+
+            # Leg resistivity (different leg temperature as separate cooling channels)
+            if tfcoil_variables.i_tf_sup == 0:
+                tfcoil_variables.rho_tf_leg = (
+                    tfcoil_variables.frholeg
+                    * (
+                        1.86e0
+                        + 0.00393e0
+                        * (tfcoil_variables.temp_tf_legs_outboard - 293.15e0)
+                    )
+                    * 1.0e-8
+                )
+            elif tfcoil_variables.i_tf_sup == 2:
+                tfcoil_variables.rho_tf_leg = tfcoil_variables.frholeg * (
+                    2.00016e-14 * tfcoil_variables.temp_tf_legs_outboard**3
+                    - 6.75384e-13 * tfcoil_variables.temp_tf_legs_outboard**2
+                    + 8.89159e-12 * tfcoil_variables.temp_tf_legs_outboard
+                )
+
+            # Tricky trick to make the leg / CP tempearture the same
+            if sctfcoil_module.is_leg_cp_temp_same == 1:
+                tfcoil_variables.temp_tf_legs_outboard = -1.0e0
+
+            # Centrepost resisitivity and conductor/insulation volume
+
+            (
+                tfcoil_variables.a_cp_cool,
+                tfcoil_variables.vol_cond_cp,
+                tfcoil_variables.p_cp_resistive,
+                sctfcoil_module.vol_ins_cp,
+                sctfcoil_module.vol_case_cp,
+                sctfcoil_module.vol_gr_ins_cp,
+            ) = self.cpost(
+                build_variables.r_tf_inboard_in,
+                build_variables.r_tf_inboard_out,
+                build_variables.r_cp_top,
+                sctfcoil_module.h_cp_top,
+                build_variables.hmax + build_variables.dr_tf_outboard,
+                tfcoil_variables.thkcas,
+                tfcoil_variables.casthi,
+                tfcoil_variables.tinstf,
+                tfcoil_variables.thicndut,
+                tfcoil_variables.n_tf_turn,
+                tfcoil_variables.c_tf_total,
+                tfcoil_variables.rho_cp,
+                tfcoil_variables.fcoolcp,
+                tfcoil_variables.n_tf_coils,
+            )
+
+        # Leg cross-section areas
+        # Rem : For physics_variables.itart = 1, these quantitire corresponds to the outer leg only
+        # ---
+        # Leg ground insulation area per coil [m2]
+        sctfcoil_module.a_leg_gr_ins = tfcoil_variables.a_tf_leg_outboard - (
+            tfcoil_variables.tftort - 2.0e0 * tfcoil_variables.tinstf
+        ) * (build_variables.dr_tf_outboard - 2.0e0 * tfcoil_variables.tinstf)
+
+        # Outboard leg turns insulation area per coil [m2]
+        sctfcoil_module.a_leg_ins = 2.0e0 * tfcoil_variables.thicndut * (
+            tfcoil_variables.tftort - 2.0e0 * tfcoil_variables.tinstf
+        ) + 2.0e0 * tfcoil_variables.thicndut * tfcoil_variables.n_tf_turn * (
+            build_variables.dr_tf_outboard
+            - 2.0e0 * (tfcoil_variables.thicndut + tfcoil_variables.tinstf)
+        )  # toroidal direction + radial direction
+
+        # Exact TF outboard leg conductor area per coil [m2]
+        sctfcoil_module.a_leg_cond = (1.0e0 - tfcoil_variables.f_a_tf_cool_outboard) * (
+            tfcoil_variables.a_tf_leg_outboard
+            - sctfcoil_module.a_leg_gr_ins
+            - sctfcoil_module.a_leg_ins
+        )
+        # ---
+
+        if physics_variables.itart == 1:
+            # Outer leg resistive power loss
+            # ---
+            # TF outboard leg's resistance calculation (per leg) [ohm]
+            tfcoil_variables.res_tf_leg = (
+                tfcoil_variables.rho_tf_leg
+                * tfcoil_variables.len_tf_coil
+                / sctfcoil_module.a_leg_cond
+            )
+
+            # TF outer leg resistive power (TOTAL) [W]
+            tfcoil_variables.p_tf_leg_resistive = (
+                tfcoil_variables.res_tf_leg
+                * (tfcoil_variables.c_tf_total / tfcoil_variables.n_tf_coils) ** 2
+            ) * tfcoil_variables.n_tf_coils
+            # ---
+
+            # Sliding joints resistive heating
+            # ---
+            if tfcoil_variables.i_cp_joints != 0:
+                # Number of contact area per joint (all legs)
+                n_contact_tot = (
+                    tfcoil_variables.n_tf_joints_contact
+                    * np.round(tfcoil_variables.n_tf_turn)
+                    * np.round(tfcoil_variables.n_tf_coils)
+                )
+
+                # Area of joint contact (all legs)
+                a_joints = (
+                    build_variables.dr_tf_outboard
+                    * tfcoil_variables.th_joint_contact
+                    * n_contact_tot
+                )
+
+                # Total joints resistive power losses
+                tfcoil_variables.pres_joints = (
+                    tfcoil_variables.n_tf_joints
+                    * tfcoil_variables.rho_tf_joints
+                    * tfcoil_variables.c_tf_total**2
+                    / a_joints
+                )
+            else:
+                # Joints resistance to be evaluated for SC
+                tfcoil_variables.pres_joints = 0.0e0
+
+            # ---
+
+        # Case of a resistive magnet without joints
+        # ***
+        else:
+            # TF resistive powers
+            tfcoil_variables.p_cp_resistive = (
+                tfcoil_variables.rho_cp
+                * tfcoil_variables.c_tf_total**2
+                * tfcoil_variables.len_tf_coil
+                / (sctfcoil_module.a_leg_cond * tfcoil_variables.n_tf_coils)
+            )
+
+            # tfcoil_variables.p_cp_resistive containts the the total resistive power losses
+            tfcoil_variables.p_tf_leg_resistive = 0.0e0
+
+            # No joints if physics_variables.itart = 0
+            tfcoil_variables.pres_joints = 0.0e0
+
+    @staticmethod
+    @numba.njit(cache=True)
+    def cpost(
+        r_tf_inboard_in,
+        r_tf_inboard_out,
+        r_cp_top,
+        ztop,
+        hmaxi,
+        cas_in_th,
+        cas_out_th,
+        gr_ins_th,
+        ins_th,
+        n_tf_turn,
+        curr,
+        rho,
+        fcool,
+        n_tf_coils,
+    ):
+        """
+        author: P J Knight, CCFE, Culham Science Centre
+        Calculates the volume and resistive power losses of a TART centrepost
+        This routine calculates the volume and resistive power losses
+        of a TART centrepost. It is assumed to be tapered - narrowest at
+        the midplane and reaching maximum thickness at the height of the
+        plasma. Above/below the plasma, the centrepost is cylindrical.
+        The shape of the taper is assumed to be an arc of a circle.
+        P J Knight, CCFE, Culham Science Centre
+        21/10/96 PJK Initial version
+        08/05/12 PJK Initial F90 version
+        16/10/12 PJK Added constants; removed argument pi
+        26/06/14 PJK Added error handling
+        12/11/19 SK Using fixed cooling cross-section area along the CP
+        26/11/19 SK added the coolant area, the conuctor/isulator/outer casing volume
+        30/11/20 SK added the ground outer ground insulation volume
+        F/MI/PJK/LOGBOOK12, pp.33,34
+        """
+        yy_ins = np.zeros((101,))  # Exact conductor area (to be integrated)
+        yy_cond = np.zeros((101,))  # Turn insulation area (to be integrated)
+        yy_gr_ins = np.zeros((101,))  # Outter ground insulation area (to be integrated)
+        yy_casout = np.zeros((101,))  # Outter case area (to be integrated)
+
+        rtop = r_cp_top - cas_out_th - gr_ins_th
+
+        # Conductor outer radius at CP mid-plane [m]
+        rmid = r_tf_inboard_out - cas_out_th - gr_ins_th
+
+        # Conductor inner radius [m]
+        r_tfin_inleg = r_tf_inboard_in + cas_in_th + gr_ins_th
+        # -#
+
+        #  Error traps
+        # ------------
+        # if rtop <= 0.0e0:
+        #     error_handling.fdiags[0] = rtop
+        #     error_handling.report_error(115)
+
+        # if ztop <= 0.0e0:
+        #     error_handling.fdiags[0] = ztop
+        #     error_handling.report_error(116)
+
+        # if rmid <= 0.0e0:
+        #     error_handling.fdiags[0] = rmid
+        #     error_handling.report_error(117)
+
+        # if build_variables.hmax <= 0.0e0:
+        #     error_handling.fdiags[0] = build_variables.hmax
+        #     error_handling.report_error(118)
+
+        # if (fcool < 0.0e0) or (fcool > 1.0e0):
+        #     error_handling.fdiags[0] = fcool
+        #     error_handling.report_error(119)
+
+        # if rtop < rmid:
+        #     error_handling.fdiags[0] = rtop
+        #     error_handling.fdiags[1] = rmid
+        #     error_handling.report_error(120)
+
+        # if build_variables.hmax < ztop:
+        #     error_handling.fdiags[0] = build_variables.hmax
+        #     error_handling.fdiags[1] = ztop
+        #     error_handling.report_error(121)
+
+        # ------------
+
+        # Mid-plane area calculations
+        # ---------------------------
+        # Total number of CP turns
+        n_turns_tot = n_tf_coils * n_tf_turn
+
+        # Area of the innner TF central hole [m2]
+        a_tfin_hole = np.pi * r_tfin_inleg**2
+
+        # Mid-plane outer casing cross-section area [m2]
+        a_casout = np.pi * (
+            (rmid + gr_ins_th + cas_out_th) ** 2 - (rmid + gr_ins_th) ** 2
+        )
+
+        # Mid-plane outter ground insulation thickness [m2]
+        a_cp_gr_ins = (
+            np.pi * ((rmid + gr_ins_th) ** 2 - rmid**2)
+            + 2.0e0 * gr_ins_th * (rmid - r_tfin_inleg) * n_tf_coils
+        )
+
+        # Mid-plane turn layer cross-section area [m2]
+        a_cp_ins = (
+            np.pi
+            * ((r_tfin_inleg + ins_th) ** 2 - r_tfin_inleg**2)  # Inner layer volume
+            + np.pi * (rmid**2 - (rmid - ins_th) ** 2)  # Outter layer volume
+            + 2.0e0 * n_turns_tot * ins_th * (rmid - r_tfin_inleg - 2.0e0 * ins_th)
+        )  # inter turn separtion
+
+        # Cooling pipes cross-section per coil [m2]
+        a_cp_cool = fcool * (
+            (np.pi * rmid**2 - a_tfin_hole - a_cp_ins) / n_tf_coils
+            - 2.0e0 * gr_ins_th * (rmid - r_tfin_inleg)
+        )  # Wedge ground insulation
+        # ---------------------------
+
+        #  Trivial solutions
+        # ------------------
+        if np.abs(fcool) < EPS:
+            vol_cond_cp = 0.0e0
+            respow = 0.0e0
+            vol_case_cp = 0.0e0
+            vol_gr_ins_cp = 0.0e0
+            vol_ins_cp = 0.0e0
+            # error_handling.report_error(122)
+            return (
+                a_cp_cool,
+                vol_cond_cp,
+                respow,
+                vol_ins_cp,
+                vol_case_cp,
+                vol_gr_ins_cp,
+            )
+
+        if np.abs(rmid - rtop) < EPS:
+            # Exact conductor cross-section
+            a_cond_midplane = (
+                np.pi * rmid**2 - a_tfin_hole - n_tf_coils * a_cp_cool - a_cp_ins
+            )
+
+            # Volumes and resisitive losses calculations
+            vol_cond_cp = 2.0e0 * hmaxi * a_cond_midplane
+            vol_ins_cp = 2.0e0 * hmaxi * a_cp_ins
+            vol_gr_ins_cp = 2.0e0 * hmaxi * a_cp_gr_ins
+            respow = 2.0e0 * hmaxi * curr**2 * rho / a_cond_midplane
+            vol_case_cp = 2.0e0 * hmaxi * a_casout
+
+            return (
+                a_cp_cool,
+                vol_cond_cp,
+                respow,
+                vol_ins_cp,
+                vol_case_cp,
+                vol_gr_ins_cp,
+            )
+
+        # ------------------
+
+        # Find centre of circle (RC,0) defining the taper's arc
+        # (r1,z1) is midpoint of line joining (rmid,0) and (rtop,ztop)
+        # Rem : The taper arc is defined using the outer radius of the
+        #       conductor including turn unsulation
+        # -------------------------------------------------------------
+        r1 = 0.5e0 * (rmid + rtop)
+        z1 = 0.5e0 * ztop
+
+        x = (r1 - rmid) ** 2 + z1**2
+        y = ztop**2 / ((rtop - rmid) ** 2 + ztop**2)
+
+        rc = rmid + np.sqrt(x / (1.0e0 - y))
+        # -------------------------------------------------------------
+
+        #  Find volume of tapered section of centrepost, and the resistive
+        #  power losses, by integrating along the centrepost from the midplane
+        # --------------------------------------------------------------------
+        #  Calculate centrepost radius and cross-sectional areas at each Z
+        dz = 0.01e0 * ztop
+
+        for ii in range(101):
+            z = ii * dz
+            z = np.fmin(np.array(z), ztop)
+
+            r = rc - np.sqrt((rc - rmid) ** 2 - z * z)
+
+            # if r <= 0.0e0:
+            #     error_handling.fdiags[0] = r
+            #     error_handling.fdiags[1] = rc
+            #     error_handling.fdiags[2] = rmid
+            #     error_handling.fdiags[3] = z
+
+            #     error_handling.report_error(123)
+
+            # Insulation cross-sectional area at z
+            yy_ins[ii] = (
+                np.pi * ((r_tfin_inleg + ins_th) ** 2 - r_tfin_inleg**2)
+                + np.pi * (r**2 - (r - ins_th) ** 2)  # Inner layer volume
+                + 2.0e0  # Outter layer volume
+                * ins_th
+                * (r - r_tfin_inleg - 2.0e0 * ins_th)
+                * n_turns_tot
+            )  # inter turn layers
+
+            #  Conductor cross-sectional area at z
+            yy_cond[ii] = (
+                np.pi * r**2
+                - a_tfin_hole
+                - n_tf_coils * a_cp_cool
+                - yy_ins[ii]
+                - 2.0e0 * n_tf_coils * gr_ins_th * (r - r_tfin_inleg)
+            )  # Wedge ground insulation
+
+            #  Outer ground insulation area at z
+            yy_gr_ins[ii] = np.pi * (
+                (r + gr_ins_th) ** 2 - r**2
+            ) + 2.0e0 * n_tf_coils * gr_ins_th * (r - r_tfin_inleg)
+
+            #  Outer casing Cross-sectional area at z
+            yy_casout[ii] = np.pi * (
+                (r + gr_ins_th + cas_out_th) ** 2 - (r + gr_ins_th) ** 2
+            )
+
+        #  Perform integrals using trapezium rule
+        sum1 = 0.0e0
+        sum2 = 0.0e0
+        sum3 = 0.0e0
+        sum4 = 0.0e0
+        sum5 = 0.0e0
+        for ii in range(1, 100):
+            sum1 = sum1 + yy_cond[ii]
+            sum2 = sum2 + 1.0e0 / yy_cond[ii]
+            sum3 = sum3 + yy_ins[ii]
+            sum4 = sum4 + yy_casout[ii]
+            sum5 = sum5 + yy_gr_ins[ii]
+
+        sum1 = 0.5e0 * dz * (yy_cond[0] + yy_cond[100] + 2.0e0 * sum1)
+        sum2 = 0.5e0 * dz * (1.0e0 / yy_cond[0] + 1.0e0 / yy_cond[100] + 2.0e0 * sum2)
+        sum3 = 0.5e0 * dz * (yy_ins[0] + yy_ins[100] + 2.0e0 * sum3)
+        sum4 = 0.5e0 * dz * (yy_casout[0] + yy_casout[100] + 2.0e0 * sum4)
+        sum5 = 0.5e0 * dz * (yy_gr_ins[0] + yy_gr_ins[100] + 2.0e0 * sum5)
+
+        # Turn insulation layer cross section at CP top  [m2]
+        a_cp_ins = (
+            np.pi * ((r_tfin_inleg + ins_th) ** 2 - r_tfin_inleg**2)
+            + np.pi * (rtop**2 - (rtop - ins_th) ** 2)  # Inner layer volume
+            + 2.0e0  # Outter layer volume
+            * ins_th
+            * (rtop - r_tfin_inleg - 2.0e0 * ins_th)
+            * n_turns_tot
+        )  # turn separtion layers
+
+        # Ground insulation layer cross-section at CP top [m2]
+        a_cp_gr_ins = (
+            np.pi * ((rtop + gr_ins_th) ** 2 - rtop**2)
+            + 2.0e0 * gr_ins_th * (rtop - r_tfin_inleg) * n_tf_coils
+        )
+
+        # Outer casing cross-section area at CP top [m2]
+        a_casout = np.pi * (
+            (rmid + gr_ins_th + cas_out_th) ** 2 - (rmid + gr_ins_th) ** 2
+        )
+
+        # Centrepost volume (ignoring coolant fraction) [m3]
+        vol_cond_cp = 2.0e0 * sum1 + 2.0e0 * (  # Tapered section
+            hmaxi - ztop
+        ) * (  # Straight section vertical height
+            np.pi * rtop**2
+            - a_tfin_hole
+            - a_cp_ins
+            - n_tf_coils * a_cp_cool
+            - 2.0e0 * n_tf_coils * gr_ins_th * (rtop - r_tfin_inleg)
+        )  # subtracting ground insulation wedge separation
+
+        # Resistive power losses in taped section (variable radius section) [W]
+        res_taped = rho * curr**2 * sum2
+
+        # Centrepost insulator volume [m3]
+        vol_ins_cp = 2.0e0 * (sum3 + (hmaxi - ztop) * a_cp_ins)
+
+        # Ground insulation volume [m3]
+        vol_gr_ins_cp = 2.0e0 * (
+            sum5
+            + (hmaxi - ztop) * a_cp_gr_ins
+            + hmaxi * np.pi * (r_tfin_inleg**2 - (r_tfin_inleg - gr_ins_th) ** 2)
+        )
+
+        # CP casing volume [m3]
+        vol_case_cp = 2.0e0 * (
+            sum4
+            + (hmaxi - ztop) * a_casout
+            + hmaxi
+            * np.pi
+            * (
+                (r_tfin_inleg - gr_ins_th) ** 2
+                - (r_tfin_inleg - gr_ins_th - cas_in_th) ** 2
+            )
+        )
+
+        # Resistive power losses in cylindrical section (constant radius) [W]
+        res_cyl = (
+            rho
+            * curr**2
+            * (
+                (hmaxi - ztop)
+                / (
+                    np.pi * rtop**2
+                    - a_tfin_hole
+                    - a_cp_ins
+                    - n_tf_coils * a_cp_cool
+                    - 2.0e0 * n_tf_coils * gr_ins_th * (rtop - r_tfin_inleg)
+                )
+            )
+        )  # ground insulation separation
+
+        # Total CP resistive power [W]
+        respow = 2.0e0 * (res_cyl + res_taped)
+
+        return (
+            a_cp_cool,
+            vol_cond_cp,
+            respow,
+            vol_ins_cp,
+            vol_case_cp,
+            vol_gr_ins_cp,
+        )
